integrations/pydanticai/pyproject.toml
[tool.poetry]
name = "example_server"
version = "0.1.0"
description = ""
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ag-ui-protocol = "^0.1.5"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
google-genai = "^1.27.0"
python-dotenv = "^1.1.1"
openai = "^1.98.0"
pydantic-settings = "^2.10.1"
pydantic-ai-slim = {extras = ["google"], version = "^0.8.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "example_server:main"